[lotus] Support lotus transaction merkle proofs
Lotus merkle roots include commitments to both the TxId and the TxHash
due to the inclusion of MalFix. We also fixed the CVE regarding duplicate
transactions causing a potential vulnerability. This commit updates the
Fulcrum codebase to store the tx commitment Hash(txid||txhash) as well
as using a null hash when the number of elements is odd in the merkle
tree.